Add: debug and prod server
