+++ main.py
@@ -1,0 +1,6 @@
+from viz import app
+from viz.index import *
+application = app.server
+
+if __name__ == '__main__':
+    app.run_server(debug=True)
